Use `dep` to manage dependencies
Currently the image inspector uses the `trash` tool to manage the
dependencies, using the `vendor.conf` file to declare them, and
commiting `vendor` directory to the source code repository.

This patch changes the project to use the `dep` tool and removes the
`vendor` directory from the source code repository. The dependencies
will be downloaded by the `dep` tool when running `make`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>