Rearrange Parameters in readOpenSCAPReports()
I believe that the parameters in `readOpenSCAPReports()` have been incorrectly placed. When running the `image-inspector` command with `--openscap-html-report` specified, it currently tries to run `xmldom.ParseXML(string(report))` (line 264) on the resulting html file. This results in a parse error that I was able to resolve by switching the order of the two parameters on line 253